refactor(react-gui): separate the animation inspector's form and its generic tab
AnimElementInspector was 742 lines holding three things: the conversions
between an element's stored numbers and the text fields that edit them,
the Generic tab's raw property table, and the typed editor itself.

animElementForm.ts (74 lines, no React) holds the first. The editor keeps
its own FormState rather than editing the fetched detail in place, because
the fields are text while the element stores numbers -- a half-typed value
has to survive on screen without being written -- so these conversions are
the seam between the two, and they are worth testing.

Testing them corrected me. `wrapAngle(720)` returns 360, not 0: the loop
is `> 360`, not `>= 360`, so a full turn stays a full turn. That is not an
oversight -- dragging up through the top of the range reads continuously
instead of snapping back to zero, and 360 and 0 are the same rotation. The
test pins the behaviour that exists and says why, rather than the one I
assumed. The other five cover the axis presets (a scaled axis is NOT the
preset, or the dropdown would misreport the vector as unit) and the
four-decimal rounding that keeps a drag's IEEE-754 noise out of the field.

useAnimGenericProps.ts holds the second, and states the thing that looks
like a bug until you know it: the generic table is fetched in BOTH modes,
because the mode bar's Reset-all button derives its enabled state from
those entries and so needs them live on the Properties tab too.

Five exhaustive-deps omissions surfaced, all the same shape as the last
three splits: the element-identity refs were closure captures and are now
arguments, so the callbacks that read them have to say so. Ref objects are
stable, so naming them keeps every callback reference-stable -- which is
why they are refs.
